added auth/oauth2.ts added oauth2 to ProviderEnum
